fix(policy): compute teacher probe norms in float64
float32 norm accumulation truncates on multi-million-element tensors
(CPU sequential reduction underestimates a 16M-element embedding norm
by ~0.2% while GPU tree reduction stays accurate), causing spurious
teacher-vs-actor mismatches on bit-identical weights. Unblocks all 4
teacher EMA tests (job 2682034: 4 passed).

+++ nemo_rl/models/policy/workers/dtensor_policy_worker_v2.py
@@ -1876,12 +1876,17 @@ class DTensorPolicyWorkerV2(AbstractPolicyWorker, ColocatablePolicyInterface):
         if self.teacher_state_dict is None:
             return {"teacher_keys": [], "teacher_norms": {}, "actor_norms": {}, "base_weight_norms": {}}
         state = self.model.state_dict()
+        # Norms in float64: float32 accumulation over multi-million-element
+        # tensors truncates small terms once the running sum grows (CPU
+        # sequential/vectorized reduction underestimates by ~0.2% on a 16M
+        # element embedding), while GPU tree reduction does not — so float32
+        # CPU-vs-GPU norms of bit-identical tensors disagree.
         teacher_norms = {
-            k: t.detach().float().norm().item()
+            k: t.detach().double().norm().item()
             for k, t in self.teacher_state_dict.items()
         }
         actor_norms = {
-            k: to_local_if_dtensor(state[k]).detach().float().norm().item()
+            k: to_local_if_dtensor(state[k]).detach().double().norm().item()
             for k in self.teacher_state_dict
         }
         # For backward compat with tests: base_weight_norms = non-LoRA subset of teacher_norms
